Move static sets to their own module
